feat: migrated the project to typescript based
